sample/server/requirements.txt
Django==2.1.5
djangorestframework==3.9.1
Markdown==3.0.1
gunicorn==19.9.0
#pylint==2.2.2
djangorestframework-jwt==1.11.0
PyYAML==3.13
PyMySQL==0.9.3